requirements-lambda.txt
# Lambda dependencies for CTI Scraper
# Minimal set for scraping functionality

# AWS SDK (included in Lambda runtime, but pinning for consistency)
boto3>=1.34.69

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic>=1.13.0
pgvector>=0.2.4

# Web scraping
feedparser==6.0.11
beautifulsoup4==4.12.3
lxml==5.1.0
aiohttp==3.9.3
chardet==5.2.0

# Utilities
python-dateutil==2.8.2
tenacity==8.2.3
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.1

# ML (Phase 2) - Heavy dependencies removed for Lambda size limits
# scikit-learn, scipy removed - using HuntScorer fallback in Lambda
# ML model inference available via API endpoints with larger compute
